work around weird libyaml-cpp bug
Somehow at the last element of the main entry sequence, dereferencing an
iterator of a sub-sequence leads to the toplevel node being reset. Don't
have the nerve to track down the real issue right now, but if we just
copy the toplevel node, it doesn't get overwritten. What a mess.