Implement general driver initialization retries
This comes with some refactoring in the way the runtime temperature
state is referenced and updated.

The main behavioral change however is that all drivers now have a
max_errors setting and thinkfan will fail only when some driver has
exceeded that maximum number of (consecutive!) errors. In addition, fans
can also be marked optional now.

Note that this is only the core functionality. Some cleanup and the
necessary config additions are yet to come.