factor out libsensors interface
We need to have a singleton libsensors interface because libsensors
imposes a shared state that is coupled to the the sensors available at
initialization. So to be able to detect newly appearing sensors, we need
to shutdown the entire libsensors interface and re-initialize it, in the
process destroying the state of all previously initialized sensors,
including those that may have already been in use.